Move extraction of .data from event payloads to Listeners#notify
Before this, event payloads were handled differently depending on
whether the listener was registered with `.on()` or `.addListeners()`.
To ease migration to the new method, the behavior should be the same.
This is also the behavior we want, as we don't have a use case for
returning the event source the callbacks.