Allow passing payloads without a data key the listeners
We've been somewhat inconsistent with how payloads from messages are
handled. Specifically, whether or not we use the data key or not.

We need to revisit and document the expected format different places,
but for now, this will at least make events that do not have a .data
key work.